Refactor Azure figure saving functionality
* Moved `save_figure_with_azure_copy` function to a new module `corebehrt.functional.utils.azure_save` for better organization.
* Updated all references to the function across various modules to ensure consistent usage.
* Removed the original function implementation from the Azure utility module to streamline the codebase.
