+++ corebehrt/azure/util/__init__.py
@@ -1,17 +1,4 @@
-from corebehrt.azure.util.azure import (
-    is_azure_available,
-    check_azure,
-    ml_client,
-    save_figure_with_azure_copy,
-)
+from corebehrt.azure.util.azure import is_azure_available, check_azure, ml_client
 from corebehrt.azure.util import job, pipeline, test
 
-__all__ = [
-    is_azure_available,
-    check_azure,
-    ml_client,
-    save_figure_with_azure_copy,
-    job,
-    pipeline,
-    test,
-]
+__all__ = [is_azure_available, check_azure, ml_client, job, pipeline, test]
